Move constants and messages to variables folder
