+++ models/PokemonFight.js
@@ -1,7 +1,7 @@
 const mongoose = require('mongoose');
 const Schema = mongoose.Schema;
 
-const { msgInvalidWinner } = require('../messages.js');
+const { msgInvalidWinner } = require('../variables/messages.js');
 
 const pokemonFightSchema = new Schema({
   date: { type: Date, default: () => Date.now() },
