+++ variables/messages.js
